[DependencyInjection] Introduce a new ParameterResolver
